MAR-822: start a test notification for this functionality
minor refactor of interactions/models.py to make testing easier
